Introduce a settings DSL for central dependency declaration
This commit introduces a new DSL, available on `Settings` via the
dependency resolution management block, which allows declaring aliases
for dependencies and bundles of dependencies.

Doing this, Gradle will automatically generate type-safe accessors
which are exposed as extensions to all projects. This effectively
allows sharing dependency declarations between projects.

Said differently, this is an officially supported pattern for the
various "dependency version sharing" patterns that we found in the
wild:

- declaring dependency versions in gradle.properties, then referencing
the version via `project.someLibraryVersion`
- creating a new class in `buildSrc` which contains dependency coordinates
and then can be accessed in the different projects
- creating extra properties for dependency versions at the root project
- using external plugins like the "refresh dependencies" plugin
- ...

In addition to this DSL, we introduce a TOML file, which, if found under
the `gradle` directory, will be used to source dependency aliases.

For example, giving the following file:

```
[dependencies]
guava = { group = "com.google.guava", name = "guava", version = "27.0-jre" }
groovy = { group = "org.codehaus.groovy", name = "groovy", version.strictly = "3.0.2" }
```

a dependency can be added in a project using the type-safe `libs` extension:

```
dependencies {
    api(libs.guava)
}
```

This file also supports bundles of dependencies, in case more than one
dependency needs to be added:

```
[bundles]
groovy = ["groovy", "groovy-json"]
```

then dependencies can be added via:

```
dependencies {
   implementation(libs.groovyBundle)
}
```

The file is configuration-cache safe: any change to the file will invalidate the
cache. But more importantly, it's build cache safe: if the aliases don't change,
there's no need to rebuild the dependent scripts. This means that changing this
file by changing, for example, the dependency versions, will not trigger a
recompilation of build scripts like some of the approaches described above.

In addition, we also generate type-safe accessors for projects. For example:

```
dependencies {
    api(project(":commons:core"))
}
```

can be replaced with:

```
dependencies {
    api(projects.commons.core)
}
```

Therefore any change to project coordinates, or removals of subprojects would
trigger a build script compilation error, avoiding tedious search and replace.